+++ src/components/image-manager/ServerImagePreview.tsx
@@ -1,23 +1,25 @@
 import { Box, Button, Card, CardMedia } from "@mui/material";
 import DeleteIcon from "@mui/icons-material/Delete";
 
-export default function Preview(serverImage: string, path: string) {
-  const handleDelete = () => {};
+interface ServerImagePreview {
+  onImageDelete: () => void;
+  serverImage: string;
+}
 
+export default function ServerImagePreview({
+  onImageDelete,
+  serverImage,
+}: ServerImagePreview) {
   return (
     <Card>
-      <CardMedia
-        component="img"
-        image={serverImage}
-        alt="Server content"
-        sx={{ maxHeight: 400 }}
-      />
+      <CardMedia component="img" image={serverImage} alt="Server content" />
       <Box sx={{ display: "flex", justifyContent: "center", p: 2 }}>
         <Button
           variant="contained"
+          size="small"
           color="error"
           startIcon={<DeleteIcon />}
-          onClick={handleDelete}
+          onClick={onImageDelete}
         >
           Obriši
         </Button>
